requirements.txt
slack-bolt
openai
python-dotenv
apscheduler
requests
